sp-42: update word
